requirement.txt
opencv-python
scikit-image
pillow
torchvision
hdf5storage
ninja
mmcv>=0.4.4
numpy
torch>=1.1